+++ src/components/autocompleter/autocompleter.test.jsx
@@ -28,6 +28,20 @@ describe('src/components/autocompleter/autocompleter', () => {
     expect(ref.current).toMatchSnapshot();
   });
 
+  describe('apiEndpoint API', () => {
+    it('fetches', async () => {
+      const { rerender } = render(<Autocompleter {...requiredProps} apiEndpoint={'/models?only=55'} />);
+      userEvent.click(screen.getByLabelText('Test label'));
+      const foo = await screen.findByText('Foo');
+      expect(foo).toBeInTheDocument();
+
+      rerender(<Autocompleter {...requiredProps} apiEndpoint={'/models?only=9'} />);
+      const option9 = await screen.findByText('Option 9');
+      expect(option9).toBeInTheDocument();
+      expect(foo).not.toBeInTheDocument();
+    });
+  });
+
   describe('className API', () => {
     it('prioritizes className prop', () => {
       const { container } = render(<Autocompleter {...requiredProps} className="prioritize-me" />);